pyproject.toml
[tool.poetry]
name = "ai-readme-generation"
version = "0.1.0"
description = "A tool for automatically generating README files for code repositories using LLM analysis."
authors = ["AI4DP"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "*"
black = "*"
langchain = "*"
langchain-openai = "*"
openai = "*"

[tool.poetry.group.dev.dependencies]
# Development dependencies can be added here

[tool.poetry.group.test.dependencies]
# Test dependencies can be added here

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
